perf(dsv4-trtllm): emit fp8 q directly from the fused q norm+rope kernel
The trtllm-gen backend consumes q as e4m3, but fused_q_norm_rope stored
bf16 and the backend then ran a separate per-layer cast pass
(ConvertToFloat8E4M3fnOp, ~0.5 percent of steady-state kernel time and
one full extra read+write of q per layer). Add a kFp8Out variant that
does the e4m3 pack at the store -- bit-identical to the bf16 store
followed by .to(float8_e4m3fn), since the bf16 rounding already happens
in-kernel. The variant is selected by allocating q_out as e4m3
(_compute_q_b does this when the attention backend is trtllm and no
preallocated bf16 q_out is passed); the backend skips its cast when q
already arrives as e4m3. Output buffers previously dtype-derived from q
(idle path, BCG prefill o) are pinned to bf16.

Unit test: fp8-out vs bf16-out + cast, bitwise equal across token/head
shapes.

+++ python/sglang/srt/layers/attention/deepseek_v4_backend.py
@@ -2021,7 +2021,11 @@ class DeepseekV4AttnBackend(
         **_,
     ) -> torch.Tensor:
         if self.mtp_enabled and forward_batch.forward_mode.is_idle():
-            return q.new_empty(q.shape[0], q.shape[1], layer.v_head_dim)
+            # Attention emits bf16 regardless of q's dtype (q may be e4m3 on
+            # the trtllm fused-q path); don't derive the output dtype from q.
+            return q.new_empty(
+                q.shape[0], q.shape[1], layer.v_head_dim, dtype=torch.bfloat16
+            )
 
         assert k is v, "DeepseekV4 shares k and v"
         swa_k = k
